Refactored to egm.
